refactor: move middlewares into separate package
